fix(HCNetSDK): 修改记录通道数据类型为IntPtr以提高内存管理 fix(manifest): 更新ControlEntradaSalida.exe的大小和摘要值

+++ HCNetSDK.cs
@@ -3308,7 +3308,7 @@ namespace ControlEntradaSalida
             public byte byChannelControllerIRAdaptorID;
             public byte byChannelControllerIREmitterID;
             public uint dwRecordChannelNum;
-            public uint pRecordChannelData;
+            public IntPtr pRecordChannelData;
             public byte byUserType;
             public byte byCurrentVerifyMode;
             [MarshalAs(UnmanagedType.ByValArray, SizeConst = 2)]
